add support for linux x64 (#64)
Fixes #43 

Some small code adjustments had to be made to compile the workspace with
`g++`:
- including some more headers
- removing the super-class for `enum`s
- the `std::c32rtomb` function used in `icecream.hpp` that doesn't seem
to exist in the `std` namespace (???)

There was also a much bigger issue involving `blasfeo`. Because of the
way the `conda` environment automatically edits the `RPATH` of
executables, the `libblasfeo.so` that was actually loaded by executables
like `sim_node` was the one installed in the `conda` env as a transitive
dependency of `casadi`, and not the one we compiled ourselves when
building `acados`.

On other platforms (`osx-arm64` and `linux-aarch64`), `libblasfeo` was
installed with version 0.1.4.1, and on `linux-64` an old 0.1.3 version
was installed because of some hard to resolve dependencies compatibility
issues between all the CPU architectures supported by `libblasfeo`. This
version unfortunately is missing some important symbols (e.g.
`blafeo_dsyr2k`) that are needed in `acados`.

After unsuccessfully trying to find a way to edit the `RPATH` of
`sim_node` in such a way that it loads the `libblasfeo.so` that we
compiled ourselves, we decided to instead tweak the `env.yml` to make
sure `libblasfeo>=0.1.4.1` was installed everywhere. Modulo [some bugs
in `conda-lock`](https://github.com/conda/conda-lock/pull/776), we were
able to successfully lock the environment for all three platforms
(`linux-64`, `linux-aarch64` and `osx-arm64`) by adding a
`virtual-packages.yml`. This file specifies assumptions that we make on
the CPU capabilities of the machines on which we install the
environment. In particular, we were able to specify that the Linux
x86_64 machines will support AVX2 instructions (corresponding to
archspec `x86_64_v3`), which guarantees `libblasfeo>=0.1.4.1`.