add build CI workflow (#65)
Fixes #20

This PR introduces a simple Github Actions config `build.yml` that can
either be triggered manually from the actions tab of the repo, or by
adding the `run_build_ci` label on a PR.
It runs two jobs:
- one removing the tag from the PR (to be able to trigger it again in
the future). This system allows us to only run the CI when we know to
have made meaningful changes, not every time we push something.
- one that creates a `conda` environment using the locked dependencies
in `conda-lock.yml` (using `micromamba`) and builds the workspace. The
environment is cached and should be reused when re-running the CI on the
same PR (it is however not shared between PRs and branches, see [their
docu](https://github.com/mamba-org/setup-micromamba?tab=readme-ov-file#notes-on-caching)).
This significantly reduces the setup time of the environment (from ≈1min
to ≈15s on ubuntu and from ≈5mins to ≈1min on macOS).

---------

Signed-off-by: Ted <oancea.tudor@icloud.com>