control v0 (#63)
Fixes #12, #11, #9 and #71

This PR introduces the first functional version of the control module.
It spans over 2 main parts:
- `controller.hpp` and `controller.cpp`: these files define a
`Controller` class that encompasses most of the functionality (in
particular the MPC formulation, the calls to the solvers, the initial
guess, etc.). Note that we migrated the formulation to C++ and removed
the python script because the previous interface did not let us access
the solver status.
- `control_node.cpp`: defines the `control_node` that is a wrapper
around the `Controller` class, adding the communication layer to the
other . It also publishes very basic visualization

We also made modifications to the rest of the codebase:
- added the first YAML launch files in the `launch` folder
- add more methods to the `Track` class, that now concentrates more
features related to Cartesian<->Frenet pose conversions.
- fixed issues in the track database, in particular in the center line
that was generated.
- in the `fake_track_estimation_node` we now adjust the headings that we
send to make sure that:
1. the heading of the center line on the spline interval containing the
projection of the current position is 'close to the current heading'.
Mathematically, if we call them $\varphi^\mathrm{cen}$ and $\varphi$, we
enforce that $\varphi^\mathrm{cen} \in [\varphi - \pi, \varphi + \pi]$.
This wasn't necessarily the case because we generated a heading profile
on several laps.
2. the heading values in the local spline (corresponding to the track
estimation) are continuous (no big jumps).

These adjustment will likely be useful for the actual
`track_estimation_node` in the future, so we will probably have to move
them to some other utility file.

---------

Signed-off-by: Ted <oancea.tudor@icloud.com>
Co-authored-by: Mattéo <36452374+mattbrth@users.noreply.github.com>