add new markers in control node (#90)
this PR:
- adds two additional markers for the left and right boundaries of the
track (that are tightened by the car widths and the optional inflation)
- the `control_node` also publishes on a new topic
`/brains2/control/debug_info` the reference and open-loop prediction of
the MPC (i.e. the optimal values). We wanted to include them in the
diagnostics but they need to be sent as `float64[]` instead of `string`
to be able to be visualized in foxglove, so we had to create a new
message type `ControllerDebugInfo`.
- the foxglove template now includes additional plot panels in the
`control` tab for the open-loop predictions
- the `sim_node` now publishes the cones markers every 1s to facilitate
viz in rosbags

---------

Signed-off-by: Ted <oancea.tudor@icloud.com>