add coordination node and stopping logic (#93)
fixes #77 

This PR adds a new important node, the `coordination_node`, which is in
charge to coordinate all the other components of the system. It acts as
a global FSM that knows when to start the car, when to stop it, when to
trigger an emergency, and how to inform the nodes accordingly.

---------

Signed-off-by: Ted <oancea.tudor@icloud.com>