Add 1OS steering model in HLC (#96)
this PR modifies the HLC formulation to include the same 1st Order
System (1OS) model that is used in the simulator. This gives the
controller a better idea of the delay induced by the steering actuator.
The corresponding target and actual steering values, as well as the
steering rate have also be added to the `ControllerDebugInfo` messages.

We have also implemented - but commented out - steering rate constraints
and costs. During our experiments, we have observed that adding the
constraints with the current $\dot{\delta}_\rm{max}=1 \rm{rad/s}$ makes
the car leave the track even in very simple turns (e.g. 1st or 2nd turns
on track `gamma`).
Without these constraints and costs, we observe _predicted_ steering
rate values of up to $35\rm{rad/s}$, but they are never actually
reached.
It seems like the controller benefits from "being optimistic", which
might be due to its tuning.

In light of these findings, we will close #84 for now, and investigate
in future issues the open-loop/closed-loop discrepancies and the
necessity for steering rate constraints.