Add linters and formatters (#97)
fixes #19 

this PR adds multiple checks via the
[ament_lint](https://github.com/ament/ament_lint) collection of tools:
- `ament_copyright` to check the copyright header of each file
- `ament_clang_format` to check the formatting of C/C++ files
- `ament_cpplint` to check a number of rules from the [Google C++ style
guide](https://google.github.io/styleguide/cppguide.html)

these checks are added as test that can be run via `colcon test`, and
that can therefore be checked in CI along with other unit tests. For the
moment however, we only added a `Lint` job. There is still work to do on
the unit tests.

> [!NOTE]
> We only check with these tools the files that we have actually
written. This excludes all the header only libs in
`include/brains2/external` and the code produced by `casadi` and
`acados` in `src/sim/generated`